feat(dom): fix compile warning

+++ dom/src/dom/taitank_layout_node.cc
@@ -398,7 +398,7 @@ void TaitankLayoutNode::InsertChild(std::shared_ptr<LayoutNode> child, uint32_t 
   auto node = std::static_pointer_cast<TaitankLayoutNode>(child);
   assert(node->GetLayoutEngineNodeRef() != nullptr);
   engine_node_->insertChild(node->GetLayoutEngineNodeRef(), index);
-  children_.insert(children_.begin() + index, node);
+  children_.insert(children_.begin() + static_cast<int>(index), node);
   node->parent_ = shared_from_this();
 }
